onPlaceSelectedをControllerに移動
