+++ src/Place/Models/PDB01MPlace/Contexts/PCT011MPlacesList/ModelComponents/PMC01102PlacesList/PlacesList.tsx
@@ -5,8 +5,8 @@ import { styles } from './PlacesListStyle';
 
 import { PCT011MPlacesList } from '@/Place/Models/PDB01MPlace/Contexts/PCT011MPlacesList';
 
-export const PMC01102PlacesList: FC = () => {
-	const { placesList, retrieveMore, onPlaceSelected } = useContext(PCT011MPlacesList);
+export const PMC01102PlacesList: FC = ({onPlaceSelected}) => {
+	const { placesList, retrieveMore } = useContext(PCT011MPlacesList);
 
 	return (
 		<View style={styles.list}>
