+++ src/Place/Pages/PPA001Places/Places.tsx
@@ -9,14 +9,14 @@ import { PMC01101GoogleMapPlacesList } from '@/Place/Models/PDB01MPlace/Contexts
 import { PMC01102PlacesList } from '@/Place/Models/PDB01MPlace/Contexts/PCT011MPlacesList/ModelComponents/PMC01102PlacesList/PlacesList';
 
 export const PPA001Places = ({ navigation, route }: NativeStackScreenProps<BottomTabParamList, 'PPA001Places'>) => {
-	const { onAutoCompleteClicked } = PPA001PlacesController();
+	const { onAutoCompleteClicked, onPlaceSelected } = PPA001PlacesController();
 	const { country, administrativeAreaLevel1, administrativeAreaLevel2, locality } = route.params;
 
 	return (
 		<>
 			<PMC01101GoogleMapPlacesList />
 			<PCO001SearchPlace onAutoCompleteClicked={onAutoCompleteClicked} />
-			<PMC01102PlacesList />
+			<PMC01102PlacesList onPlaceSelected={onPlaceSelected}/>
 		</>
 	);
 };
