+++ src/Place/Pages/PPA001Places/PlacesController.ts
@@ -4,9 +4,22 @@ import { GooglePlaceData, GooglePlaceDetail, AddressComponent } from 'react-nati
 import { PlacesControllerInterface } from 'spelieve-common/lib/Interfaces';
 
 import { PCT011MPlacesList } from '@/Place/Models/PDB01MPlace/Contexts/PCT011MPlacesList';
+import { useNavigation } from '@react-navigation/native';
 
 export const PPA001PlacesController = (): PlacesControllerInterface => {
-	const { setSearchedAddress, onPlaceSelected } = useContext(PCT011MPlacesList);
+	const { setSearchedAddress } = useContext(PCT011MPlacesList);
+	const navigation = useNavigation();
+
+	const onPlaceSelected = (place_id: string) => {
+		// PPA002 へ遷移
+		navigation.navigate('Place', {
+			screen: 'PPA002Place',
+			params: {
+				place_id,
+				language: 'ja',
+			},
+		});
+	};
 
 	const onAutoCompleteClicked = (data: GooglePlaceData, details: GooglePlaceDetail) => {
 		const isIncludes = (arr: Array<string>, target: Array<string>) => arr.some((el) => target.includes(el));
@@ -39,5 +52,5 @@ export const PPA001PlacesController = (): PlacesControllerInterface => {
 		}
 	};
 
-	return { onAutoCompleteClicked };
+	return { onAutoCompleteClicked, onPlaceSelected };
 };
